+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unpublished Changes
+* larva-css - Revert changes which we had done for CSS generator to use desktop font size on side skins.
 
 ## 1.27.0 09-08-2022
 * larva-patterns - Add `narrow` variant for `container` module.